No pre-commit output found.

+++ helpers/hcache_simple.py
@@ -7,10 +7,9 @@ import pickle
 import re
 from typing import Any, Callable, Dict, List, Union, cast
 
-import pandas as pd
-
 import helpers.hdbg as hdbg
 import helpers.hprint as hprint
+import helpers.hsystem as hsystem
 
 _LOG = logging.getLogger(__name__)
 
@@ -407,7 +406,7 @@ def get_cache(func_name: str) -> _CacheType:
 # #############################################################################
 
 
-def cache_stats_to_str(func_name: str = "") -> pd.DataFrame:
+def cache_stats_to_str(func_name: str = "") -> "pd.DataFrame":  # noqa: F821
     """
     Print the cache stats for a function or for all functions.
 
@@ -419,6 +418,9 @@ def cache_stats_to_str(func_name: str = "") -> pd.DataFrame:
       memory: -
       disk: 2322
     """
+    # We want to limit the dependency from pandas in the cache.
+    import pandas as pd
+
     if func_name == "":
         result = []
         for func_name in get_cache_func_names("all"):
@@ -460,20 +462,35 @@ def reset_mem_cache(func_name: str = "") -> None:
     del _CACHE[func_name]
 
 
-def reset_disk_cache(func_name: str = "") -> None:
-    assert 0
+def reset_disk_cache(func_name: str = "", interactive: bool = True) -> None:
+    """
+    Reset the disk cache for a given function name.
+
+    If `func_name` is empty, reset all disk cache files.
+    :param func_name: The name of the function whose disk cache is to
+        be reset. If empty, reset all disk cache files.
+    :param interactive: If True, prompt the user for confirmation before
+        resetting the disk cache.
+    """
+    if interactive and not func_name:
+        hsystem.query_yes_no(
+            "Are you sure you want to reset the disk cache? This will delete all cache "
+        )
     if func_name == "":
         cache_files = glob.glob("cache.*")
+        _LOG.warning("Resetting disk cache")
         for file_name in cache_files:
             os.remove(file_name)
         return
     file_name = _get_cache_file_name(func_name)
-    os.remove(file_name)
+    if os.path.exists(file_name):
+        _LOG.warning(f"Removing cache file '{file_name}'")
+        os.remove(file_name)
 
 
-def reset_cache(func_name: str = "") -> None:
+def reset_cache(func_name: str = "", interactive=True) -> None:
     reset_mem_cache(func_name)
-    reset_disk_cache(func_name)
+    reset_disk_cache(func_name, interactive=interactive)
 
 
 # #############################################################################
